Rebase main branch onto same upstream version as rename-to-mos.
To get the tests passing, moved lld-reliant tests to the lld directory,
since lld automatically builds the assembler and objcopy, but the
reverse is not true. check-all always worked, but check-llvm did not.

Git diffs between rename-to-mos and this branch will only include real
effective changes between the mysterymath llvm repo and the llvm-mos
one. These differences will be gradually resolved by porting
functionality from that branch to this one.

Once this branch contains a fully-integrated version of the two
codebases, it should be pushed to the main branch. At that point,
work can continue onward on this shared copy of LLVM.