Allow __attribute__((leaf)) on inline asm statements.
Otherwise, there's no way for a user to guarantee that an inline
assembly fragment (say, one with a JSR) does not arbitrarily call
external C functions. Without this, any function that contains inline
ASM with a JSR must be assumed to call main, making that whole section
of the programs call graph possibly recursive. This would essentially
nullify the static stack optimization, so it's essential for good 6502
codegen that inline-ASM-based calls to platform OS routines be marked
as leaves when they are.