Replace hard stack with increased CSR usage.
This deletes quite a lot of code for a pretty modest performance hit.
After this change, the compiler orchestrates all it's hard stack usage
through the CSRs. The downside is that it can no longer replace STstk
and LDstk with PHA and PLA inline, but the code that did so was the
trifecta of already huge, fragile, and woefully incomplete.