Cleanly separate logical and real instr sets.
The MC and CodeGen have very different views of what the operands of an
instruction are. By modern standards, 6502 instructions vastly
underspecify their operands; the operands are encoded into the opcodes
themselves in a way that's not exactly unusual, but rare in modern
architectures.

While there's overlap between the two, the instruction set is small, and
the removed duplication in this case doesn't seem to pay for the
uncertainty of whether or not the instruction is a pseudo. By fully
duplicating the instruction set, we gain the freedom to organize the
code-generators set along purely logical grounds, while the real
instruction set can be organized along bit-pattern grounds. TableGen
provides considerable support for both tasks, but on this architecture,
not at the same time.