Make LSB subregisters of GPR and Imag8 registers.
MachineVerifier doesn't permit COPY instructions between registers of
different sizes, which the register allocator will freely generate when
widening e.g. Cc regclass to Anyi1 regclass. The original intent was to
allow the GPRs and Imag8 registers to contain booleans, but a clearer
way to do so is to give them 1-bit subregisters tied to their LSB.

Hopefully, this approach will scale up to allow all 8 bits of GPRs and
Imag8 registers to be individually addressable. This, in combination
with LLVM's increasingly sophisticad subregister liveness tracking,
could allow the compiler to efficiently pack boolans as flags into
bytes. This would allow considerably greater memory and code density, as
a much greater number of boolean values could be resident in GPRs at a
time, ready to be manipulated with ORA, AND, EOR, and BIT.