Handle G_ANYEXT G_ZEXT, and G_TRUNC in legalizer.
These were handled in the instruction selector to ensure that redundant
combinations of these instructions would be optimized out in the
post-legalizer combiner phase. However, these instructions are treated
specially by the legalizer as "legalization artifacts," and a dedicated
combine phase is run in-line with the legalizer to remove redundant
instances. Accordingly, we're free to consider these instructions
illegal and lower them to G_INSERTs without risking emitting truncations
of extensions and similar patterns.