Teach MachineScheduler how to load A,X,Y for calls.
Generally, A needs to be loaded last, preceeded by X and Y, preceeded by
everything else. This is because A is often useful for loading X and Y,
and GPRs are often useful for loading zero page registers and stack. The
inverse is true for return values.

This adds a clause to the scheduler encoding this preference. The
original biasPhysReg() call is also restored; we actually do need to
keep physreg live ranges as short as possible, or the register allocator
chokes on large examples (printf).