Consider LSB subreg writes to clobber 8-bit reg.
Pretty much every MOS sequence that writes to a LSB (or MSB, if we end
up going that way) will clobber some portion of the rest of the
register. This change makes this explicit. Accordingly, the
zero-extension operation needs to be explicitly handled via pseudo; it
can no longer be emitted indirectly as a write to the LSB of an
already-zero register.

For now, we can reuse the zero-extension pseudo to perform any-extension
from 1-bit to 8-bit regs. We'll emit more efficient sequences later.