Use virtual registers in pseudo expansion and scavenge.
This replaces the very special-cased logic in the pseudo expansion with
regular register scavenging. This makes it easier to maintain pseudo
expansion, while improving the quality of the generated code sequences.
Most importantly, it provides a finer-grained analysis of whether or not
save and restore is required, which allows the machine verifier to pass
in more cases.