Allow storing partially-alive Imag16 regs.
The register allocator itself can generate stores of partially-alive
16-bit pointers.The various hacks we've been accumulating are
thus insufficient to fully prevent this case from occurring, and they
cannot straightforwardly be extended to do so.

Accordingly, this adds direct support for this in the stack store code
and removes all the hacks used to avoid this case. Before storing a
16-bit vreg, we use a KILL to make it fully alive.