If we have interrupts, mark all libcalls recursive.
Whether or not a function calls a libcall depends on how exactly it's
compiled; this may in turn depend on whether or not the function is
norecurse. This creates a chicken-and-egg problem, so we have to
summarily consider libcalls possibly recursive just in case an interrupt
function might call them.