Revert "Remove __save_x and __save_y."
This reverts commit 25d5e05a3ec73bf32bf08ce732f2456b140dfa5d.

This ultimately creates untenable situations in the register scavenger.
To get around this, we'd need to synthesize PHX and PHY pseudos, which
is extraordinarily expensive. (Push a dummy byte, A, and X to the stack,
save Y to the stack via TSX and indexed STY, then pop X and A, leaving Y
on the stack.). It's better just to use the original temporaries, since
the cost of saving the temporary once in an interrupt is much less than
issuing a PHY once in an interrupt.