Rewrite IV uses for non-native integer types.
A good portion of loops on MOS will have core IVs of 16-bit, and without
rewriting their uses, we don't even get the count-to-zero optimization.
It's essential that all IV uses are rewritten on MOS. However, the LSR
pass doesn't currently do a very good job at avoiding very wide illegal
IV expressions; that's why it was disabled.

Accordingly, enabling this breaks a lot of tests, including the MOS
Index IV optimization. This regression is expected and the tests updated
here; we'll have to go backwards to go forwards. It'll take some time to
upgrade LSR to be capable of dealing with the 6502 properly, and code
quality will need to regress until that time. Fixing up the other
targets is also optional; MOS first.