Sink nested G_SELECTs.
It's common for the result of one G_SELECT to only be used within the
true or false arm of another G_SELECT, for example, when emitting
multibyte comparisons or increments. In this case, computation of the
former G_SELECT can be conditioned on the latter. This can be done by
ensuring that the outer G_SELECT is lowered first, and that the inner
G_SELECT is sunk into the appropriate conditional basic block.

This allows a principled mechanism for establishing the sorts of control
flow chains used to lower a variety of common 6502 control flow idioms.
These should more-or-less fall out of a corresponding chain of G_SELECT
instructions, although some extensions to this idea may be required as
we go.