Disable div by constant -> mul optimization.
I was waiting excitedly for this to reach GlobalISel, but now that it
has, it's not actually a good idea for MOS! Long division and long
multiplicaiton are very nearly the same speed, but this optimization
turns a division into a multiplication at twice the width! On top of
that, it adds a variable shift, so it just doesn't seem worth it.