Only generate legal G_INDEX instructions.
Previously, G_INDEX needed special logic both in the legalizer and the
instruction selector, as G_INDEX might not have actually ended up in a
context where it could become and indexed addressing mode. Instead, the
legalizer now emits a regular pointer addition in all of those contexts,
so the instruction selector can be assured that any incoming G_INDEX are
legal. This simplifies the instruction selector and allows the legalizer
to further optimize the pointer additions.