Allow registers in LSR formulae to be narrower than the formula.
This helps to keep calculations in loops as narrow as possible.
This also paves the way to teach LSR about the 6502's real addressing
modes, which require combining a 16-bit base with a 8-bit register
offset. With this change, this forumla is now directly representable in
LSR, so it's just a matter of forwarding it properly to the target
description functions.