Move addressing mode selection from isel to legalizer.
The instruction selector is too late to intelligently handle addressing
modes. Addressing modes should generally be as greedy as possible, and
whatever is left over should be lowered to pointer additions. This is
most naturally done in the legalizer.

This introduces load and store generic opcodes for each addressing mode.
The legalizer will turn loads and stores of pointer additions into
these. This will allow the instruction selector to look for and attach
the appropriate addressing mode to various CISC instructions, although
for now only literal loads and stores are supported.