`.mos_addr_asciz` target directive with `R_MOS_ADDR_ASCIZ`
To ensure consistency in the BASIC header on C64, a special relocation
type and assembly directive are provided to output the value as a
decimal ASCII string.

Syntax: `.mos_addr_asciz <expression>, <num-digit-chars>`

A fixed count of null terminators is output to reserve space for an
address that would be practical for the target platform. Prior to
relocation, the string "0" is emitted to the object file.