[NFC] Determine the assignment of zero page locations.
This change does a dry-run of the whole-program zero page assignment
algorithm. No actions are taken, but zero page locations are allocated
as they would be in production.

Locations are allocated round-robin to each entry point. If a candidate
takes more than one byte, it is deferred until enough bytes have been
allocated to it. Once a location is allocated, the maximum ZP usage on
any SCC path through that function is bumped. This may make some WIP
candidates no longer viable. In that case, any previously allocated
locations are freed and reassigned.