Revert "Include tools needed for rust-mos build (#275) (#276)"
This reverts commit ca75934d2b4c29cc9cea38629d50b6c2f42d9016.
This doubles our installed size to 2GiB, which just isn't acceptible for
a 6502 compiler. We need to be a lot more targeted about what we include
in our release in general, so this change needs to figure out exactly
what must be included to build rust-mos.

We also may need to split out builds into a full and user pair, if too
much stuff is needed to build rust-mos.