Fix bug assigning register bank to stale DBG_VALUE.
Debug instructions are unique in that they can have virtual register
uses with no definitions. These are scrubbed out later when the debug
instructions are analyzed, but they need to be assigned a register bank
and register class in the meantime. Creating fake classes for i32 and
i64 allows this to proceed.