Wrap od65 and ld65 to link xo65 object files
This PR adds a complete subsystem to lld for handling XO65 object files
(ca65's output format) in the course of a normal link.

When input files are being parsed, xo65 files are detected via their
magic number and od65 is invoked to parse them, either using the system
path or the --od65-path flag. This allows the files to participate in
normal symbol resolution and for xo65 segments to be placed by the
linker.

Afterwards, ld65 is invoked (PATH or --ld65-path) to relocate the xo65
object files and and collect the resulting binary contents. The map file
is also examined to determine the final values of xo65 symbols, which
are then used to relocate ELF files.

While there is much xo65-specific logic added to LLD, the general
"linker enclave" approach is not very xo65-specific, and it could be
broken out into an effective interface. I've not done so here, for want
of a second implementation of that interface to drive its design.

ELF section names, types, and flags, are much richer than xo65's
segments, so this information is encoded into xo65 segment names via an
underscore-based escaping scheme:
  __  -> _             (underscore)
  _d  -> $             (dollar)
  _h  -> -             (hyphen)
  _p  -> .             (period)
  _tn -> SHT_NOBITS    (@nobits)
  _tp -> SHT_NOBITS    (@progbits)
  _fw -> SHF_WRITE     (w)
  _fx -> SHF_EXECINSTR (x)

Finally, this does not work with relocatable (-r) links, and this
produces a link error.