Make IV canonicalization prefer narrow IVs on MOS
Previously, the induction variable simplification pass (indvars) would
join up a sequence of congruent induction variables of different sizes
with the widest type. All narrower uses would then be obtained by
truncation. This is the exact opposite transformation we want on the
6502; induction variables should be simplified to the narrowest type,
then zero-extended to wider uses. This change registers the 6502's
intent for the inverse transform and makes the pass work this way when
so configured.