Fix attribute leaf bug
The GlobalAAResult scan that manages the memory effects of functions
previously relied on inline asm calling out to an external funcion in
the call graph to avoid being too presumptive about its effects.
__attribute__((leaf)) in llvm-mos breaks this call graph link, so we
need additional logic to report unknown memory effects for inline asm.

Ideally, we would be more precise, but it looks like the clobbers aren't
really processed yet, so it would be a project. For now, I'm happy
preserving the old behavior by being conservative.

Note that I did make a change to only report unknown memory effects if
the asm is volatile. Otherwise, all side effects must be the production
of outputs, and these are explicitly tracked.

Fixes #392