Strict inlining costs (#419)
The inliner will apply negative boosts to inlining costs if runtime
desirable situations occur. In particular, inlining is strongly
preferred if an indirect call can be folded away as a result.
Unfortunatley, this can blow up sizes on the 6502 arbitrarily, and
worse, it affects our printf implementation.

This change adds a new TTI hook, strictInliningCosts(), that prevents
this boost from altering the true costs. Size is paramount for our
target, so we should disable other boosts using this flag as they cause
problems.