MOS Interrupts: Add the wedge attribute
This allows us to end ISRs with a TailJMP instruction instead of
the usual RTI or RTS. This allows for ISRs that are partially
defined in the platform's ROM to be continued using a "wedge" style
interrupt mechanism, like what is typically required for Commodore
machines.

`wedge` takes an argument that is the vector to jump back to, which
is the immediate argument to the TailJMP instruction emitted in the
call lowering. `wedge` is expected to be used with the normal
`interrupt` or `interrupt_norecurse` attributes.