Remove MOSIncDecPhi pass
This pass isn't very well conceived. Issue #491 demonstrates an example,
where, if it worked properly, it would move a load into a loop
unnecessarily. I think this needs to be rewritten from the ground up,
since as is, it's a very large amount of unusually brittle and complex
code for a very small gain. Accordingly, I'm removing the pass while I
focus on the register allocator.

Fixes #491.